(tests): Fix deprecated methods calls
